Add Lane component fix game provider
