+++ src/components/Game/Lane/RowTwo.tsx
@@ -2,9 +2,9 @@ import React from 'react';
 import { StyleSheet, View } from 'react-native';
 import PropTypes from 'prop-types';
 
-import Pin from '../../../components/Pin';
+import Pin from '../../Pin';
 
-export default class Four extends React.PureComponent {
+export default class Two extends React.PureComponent {
   static propTypes = {
     pins: PropTypes.arrayOf(PropTypes.object),
     handlePinPress: PropTypes.func
@@ -17,24 +17,20 @@ export default class Four extends React.PureComponent {
 
   render() {
     const { handlePinPress, pins } = this.props;
-    const [pinSeven, pinEight, pinNine, pinTen] = pins;
+    const [pinTwo, pinThree] = pins;
     return (
       <View style={styles.container}>
+        <View style={styles.column} />
+        <View style={styles.column} />
         <View style={styles.column}>
-          <Pin pin={pinSeven} onPinPress={handlePinPress} />
+          <Pin pin={pinTwo} onPinPress={handlePinPress} />
         </View>
         <View style={styles.column} />
         <View style={styles.column}>
-          <Pin pin={pinEight} onPinPress={handlePinPress} />
+          <Pin pin={pinThree} onPinPress={handlePinPress} />
         </View>
         <View style={styles.column} />
-        <View style={styles.column}>
-          <Pin pin={pinNine} onPinPress={handlePinPress} />
-        </View>
         <View style={styles.column} />
-        <View style={styles.column}>
-          <Pin pin={pinTen} onPinPress={handlePinPress} />
-        </View>
       </View>
     );
   }
