+++ src/reducers/app.tsx
@@ -1,9 +1,4 @@
-export default function app(
-  state = {
-    description: 'Hello World'
-  },
-  action
-) {
+export default function app(state, action) {
   switch (action.type) {
     default:
       return state;
